pyproject.toml
[tool.poetry]
name = "trust4ai-bias-generator-ds"
version = "0.1.0"
description = ""
authors = ["Vicente <vicentect10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
flake8 = "^7.0.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
pydantic-settings = "^2.2.1"
mariadb = "^1.1.10"
mysqlclient = "^2.2.4"
aiomysql = "^0.2.0"
sqlmodel = "^0.0.16"
alembic = "^1.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
